# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ootils Core API",
...
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/graph": {
      "get": {
        "tags": [
          "graph"
        ],
        "summary": "Get Graph",
        "description": "Return nodes and edges for the planning graph at (item, location, scenario).",
        "operationId": "get_graph_v1_graph_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "item_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Item UUID or name",
              "title": "Item Id"
            },
            "description": "Item UUID or name"
          },
          {
            "name": "location_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Location UUID or name",
              "title": "Location Id"
            },
            "description": "Location UUID or name"
          },
          {
            "name": "depth",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5,
              "minimum": 1,
              "description": "Graph traversal depth",
              "default": 2,
              "title": "Depth"
            },
            "description": "Graph traversal depth"
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "name": "scenario_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Scenario UUID or 'baseline'",
              "title": "Scenario Id"
            },
            "description": "Scenario UUID or 'baseline'"
          },
          {
            "name": "X-Scenario-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Scenario-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/nodes": {
      "get": {
        "tags": [
          "nodes"
        ],
        "summary": "List Nodes",
        "description": "List nodes with optional filters — used for UI dropdowns (e.g. Simulate page).",
        "operationId": "list_nodes_v1_nodes_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "item_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Item Id"
            }
          },
          {
            "name": "location_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Location Id"
            }
          },
          {
            "name": "node_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Node Type"
            }
          },
          {
            "name": "scenario_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scenario Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EdgeOut": {
        "properties": {
          "edge_id": {
            "type": "string",
            "format": "uuid",
            "title": "Edge Id"
          },
          "edge_type": {
            "type": "string",
            "title": "Edge Type"
          },
          "from_node_id": {
            "type": "string",
            "format": "uuid",
            "title": "From Node Id"
          },
          "to_node_id": {
            "type": "string",
            "format": "uuid",
            "title": "To Node Id"
          },
          "priority": {
            "type": "integer",
            "title": "Priority"
          },
          "weight_ratio": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Weight Ratio"
          }
        },
        "type": "object",
        "required": [
          "edge_id",
          "edge_type",
          "from_node_id",
          "to_node_id",
          "priority",
          "weight_ratio"
        ],
        "title": "EdgeOut"
      },
...
      "GraphResponse": {
        "properties": {
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/NodeOut"
            },
            "type": "array",
            "title": "Nodes"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/EdgeOut"
            },
            "type": "array",
            "title": "Edges"
          },
          "item_id": {
            "type": "string",
            "title": "Item Id"
          },
          "location_id": {
            "type": "string",
            "title": "Location Id"
          },
          "scenario_id": {
            "type": "string",
            "format": "uuid",
            "title": "Scenario Id"
          },
          "depth": {
            "type": "integer",
            "title": "Depth"
          }
        },
        "type": "object",
        "required": [
          "nodes",
          "edges",
          "item_id",
          "location_id",
          "scenario_id",
          "depth"
        ],
        "title": "GraphResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NodeListItem": {
        "properties": {
          "node_id": {
            "type": "string",
            "format": "uuid",
            "title": "Node Id"
          },
          "node_type": {
            "type": "string",
            "title": "Node Type"
          },
          "item_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Id"
          },
          "location_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Id"
          },
          "scenario_id": {
            "type": "string",
            "format": "uuid",
            "title": "Scenario Id"
          },
          "time_ref": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Ref"
          },
          "qty": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qty"
          },
          "item_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Code"
          },
          "location_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Code"
          }
        },
        "type": "object",
        "required": [
          "node_id",
          "node_type",
          "item_id",
          "location_id",
          "scenario_id",
          "time_ref",
          "qty"
        ],
        "title": "NodeListItem"
      },
      "NodeListResponse": {
        "properties": {
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/NodeListItem"
            },
            "type": "array",
            "title": "Nodes"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "nodes",
          "total"
        ],
        "title": "NodeListResponse"
      },
      "NodeOut": {
        "properties": {
          "node_id": {
            "type": "string",
            "format": "uuid",
            "title": "Node Id"
          },
          "node_type": {
            "type": "string",
            "title": "Node Type"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "time_ref": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Ref"
          },
          "time_span_start": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Span Start"
          },
          "time_span_end": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Span End"
          },
          "time_grain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Grain"
          },
          "has_shortage": {
            "type": "boolean",
            "title": "Has Shortage"
          },
          "shortage_qty": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Shortage Qty"
          },
          "closing_stock": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closing Stock"
          }
        },
        "type": "object",
        "required": [
          "node_id",
          "node_type",
          "quantity",
          "time_ref",
          "time_span_start",
          "time_span_end",
          "time_grain",
          "has_shortage",
          "shortage_qty",
          "closing_stock"
        ],
        "title": "NodeOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}